Recursive deep filtering
Hi! 
I've noticed the stats page is still too big for nightbot because I can't include things more specifically. I've changed your function so you can further deepen your search (instead of searching multiple things).
I've tested the actual javascript with jsfiddle but I haven't tested it with node.js. Doesn't do anything crazy though so it should be good.
You could also add it as a seperate function or not at all, it's not too necessary to have (for me) but I'd like to be able to do it.

Kind regards,

Ruben
(and thanks again for your awesome API)